Flux scaling erro fix
I tracked down an error in the flux calibration stage in some data I
have been working with. If there was a nod and shuffle source observed,
then the NEXP, SEXP, NSUBEXPS header fields will still contain this
information when other observing modes are used. When the pipeline uses
the ’ns’:True parameter, signifying that N&S has been used, the exptime
is incorrectly re-calculated using these N&S parameters for exposures
made with other modes - like ClassicEquall for the standard star. This
can obviously throw off the calibration as the observed flux is scaled
in by exptime.

The fix checks the observing mode of a frame. If it is not a N&S, then
falls back to the exptime in original header, which has not been edited.